requirements.txt
pystray>=0.19.5
Pillow>=10.0.0
winotify>=1.1.0
tzdata>=2024.1
sv-ttk>=2.6.0
pyinstaller>=6.0.0
